feat!: rename prefer-create-or-replace to consistent-create-or-replace with style option
BREAKING CHANGE: `prefer-create-or-replace` is renamed to
`consistent-create-or-replace`. The new rule takes a `style` option:

- `"always"` (default): equivalent to the previous behavior — require
  `OR REPLACE` on `CREATE FUNCTION` / `PROCEDURE` / `VIEW` and auto-fix
  by inserting it.
- `"never"`: forbid `OR REPLACE` and auto-fix by removing it, so
  unintended overwrites surface as `relation already exists`.

Both stances live in one rule so they can't be enabled simultaneously
and double-report; this matches `prefer-cast-operator` / `quotes`-style
configuration.
