+++ README.md
@@ -110,6 +110,7 @@ Click a rule name to open its documentation page (examples, rationale, options).
 | ------------------------------------------------------------------------------------------------------------------------------------------------ | -------------------------------------------------------------------------- | :---------: |
 | [align-column-definitions](https://baseballyama.github.io/eslint-plugin-postgresql/rules/align-column-definitions)                               | Align column definitions vertically inside `CREATE TABLE`                  |     🎨      |
 | [align-values](https://baseballyama.github.io/eslint-plugin-postgresql/rules/align-values)                                                       | Align column values vertically inside multi-row `INSERT ... VALUES`        |     🎨      |
+| [consistent-create-or-replace](https://baseballyama.github.io/eslint-plugin-postgresql/rules/consistent-create-or-replace)                       | Enforce a consistent stance on `CREATE OR REPLACE` for FUNCTION/PROC/VIEW  |             |
 | [no-add-column-not-null-without-default](https://baseballyama.github.io/eslint-plugin-postgresql/rules/no-add-column-not-null-without-default)   | Disallow `ADD COLUMN ... NOT NULL` without a `DEFAULT`                     |     ✅      |
 | [no-alter-column-type](https://baseballyama.github.io/eslint-plugin-postgresql/rules/no-alter-column-type)                                       | Disallow `ALTER COLUMN ... TYPE` (table rewrite under exclusive lock)      |     ⚠️      |
 | [no-char-type](https://baseballyama.github.io/eslint-plugin-postgresql/rules/no-char-type)                                                       | Disallow `char(n)` / `bpchar(n)` columns                                   |     ⚠️      |
@@ -162,7 +163,6 @@ Click a rule name to open its documentation page (examples, rationale, options).
 | [prefer-cast-operator](https://baseballyama.github.io/eslint-plugin-postgresql/rules/prefer-cast-operator)                                       | Enforce a single style for type casts (`x::int` or `CAST(x AS int)`)       |     🎨      |
 | [prefer-coalesce-over-case](https://baseballyama.github.io/eslint-plugin-postgresql/rules/prefer-coalesce-over-case)                             | Prefer `COALESCE(x, y)` over a null-fallback `CASE`                        |     ⚠️      |
 | [prefer-create-index-concurrently](https://baseballyama.github.io/eslint-plugin-postgresql/rules/prefer-create-index-concurrently)               | Prefer `CREATE INDEX CONCURRENTLY`                                         |             |
-| [prefer-create-or-replace](https://baseballyama.github.io/eslint-plugin-postgresql/rules/prefer-create-or-replace)                               | Prefer `CREATE OR REPLACE` for `FUNCTION` / `PROCEDURE` / `VIEW`           |             |
 | [prefer-current-timestamp-over-now](https://baseballyama.github.io/eslint-plugin-postgresql/rules/prefer-current-timestamp-over-now)             | Prefer SQL-standard `CURRENT_TIMESTAMP` over `now()` / `LOCALTIMESTAMP`    |     🎨      |
 | [prefer-drop-index-concurrently](https://baseballyama.github.io/eslint-plugin-postgresql/rules/prefer-drop-index-concurrently)                   | Prefer `DROP INDEX CONCURRENTLY`                                           |             |
 | [prefer-exists-over-in-subquery](https://baseballyama.github.io/eslint-plugin-postgresql/rules/prefer-exists-over-in-subquery)                   | Prefer `EXISTS (subquery)` over `column IN (subquery)`                     |     ⚠️      |
