+++ tests/fixtures/consistent-create-or-replace/invalid/missing-or-replace.sql
