+++ tests/fixtures/consistent-create-or-replace/invalid/never-with-or-replace-errors.expected.json
@@ -1,0 +1,5 @@
+[
+  { "messageId": "unexpectedOrReplace" },
+  { "messageId": "unexpectedOrReplace" },
+  { "messageId": "unexpectedOrReplace" }
+]
