+++ tests/fixtures/consistent-create-or-replace/invalid/never-with-or-replace-options.json
@@ -1,0 +1,1 @@
+[{ "style": "never" }]
