+++ tests/fixtures/consistent-create-or-replace/invalid/never-with-or-replace-output.expected.sql
@@ -1,0 +1,3 @@
+CREATE FUNCTION fn() RETURNS void LANGUAGE SQL AS '';
+CREATE PROCEDURE p() LANGUAGE SQL AS '';
+CREATE VIEW v AS SELECT 1;
