+++ tests/fixtures/consistent-create-or-replace/invalid/never-with-or-replace.sql
@@ -1,0 +1,3 @@
+CREATE OR REPLACE FUNCTION fn() RETURNS void LANGUAGE SQL AS '';
+CREATE OR REPLACE PROCEDURE p() LANGUAGE SQL AS '';
+CREATE OR REPLACE VIEW v AS SELECT 1;
