+++ tests/fixtures/consistent-create-or-replace/valid/never-without-or-replace-options.json
@@ -1,0 +1,1 @@
+[{ "style": "never" }]
