+++ tests/fixtures/consistent-create-or-replace/valid/with-or-replace.sql
